refactor: pages 디렉토리 생성 및 component에서 일부 분리
