+++ fe/src/Pages/Main/Main.styled.tsx
