+++ fe/src/Pages/NewIssue/NewIssueContent/NewIssueContent.styled.tsx
