+++ fe/src/Pages/NewIssue/NewIssueContent/index.tsx
